@@ -5,13 +5,15 @@
 	<groupId>org.springframework.social</groupId>
 	<artifactId>spring-social-mailru</artifactId>
 	<name>Spring Social mail.ru</name>
-	<version>1.0-SNAPSHOT</version>
+	<version>1.1</version>
 
 	<description>Spring Social extension with connection support and an API binding for mail.ru</description>
 
 	<properties>
-		<spring.social.version>1.0.0.RELEASE</spring.social.version>
+		<spring.social.version>1.1.0.RELEASE</spring.social.version>
 	</properties>
+
+	<packaging>jar</packaging>
 
 	<repositories>
 		<repository>
@@ -47,10 +49,22 @@
 			<version>${spring.social.version}</version>
 		</dependency>
 		<dependency>
-			<groupId>org.codehaus.jackson</groupId>
-			<artifactId>jackson-mapper-asl</artifactId>
-			<version>1.8.5</version>
+			<groupId>org.springframework.social</groupId>
+			<artifactId>spring-social-config</artifactId>
+			<version>${spring.social.version}</version>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.social</groupId>
+			<artifactId>spring-social-security</artifactId>
+			<version>${spring.social.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+			<version>2.4.3</version>
+		</dependency>
+
 		<dependency>
 			<groupId>xom</groupId>
 			<artifactId>xom</artifactId>
@@ -78,13 +92,21 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>2.0.2</version>
+				<version>3.2</version>
 				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
+					<source>1.8</source>
+					<target>1.8</target>
 				</configuration>
 			</plugin>
 		</plugins>
 	</build>
+
+	<distributionManagement>
+		<repository>
+			<id>nexus-airlabs</id>
+			<name>airlabs</name>
+			<url>http://alpha.airlabs.ru:8081/nexus/content/repositories/airlabs</url>
+		</repository>
+	</distributionManagement>
 
 </project>
