@@ -15,15 +15,16 @@
  */
 package org.springframework.social.mailru.api.impl;
 
-import org.codehaus.jackson.map.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.http.converter.json.MappingJacksonHttpMessageConverter;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.social.mailru.api.Mailru;
 import org.springframework.social.mailru.api.MailruErrorHandler;
 import org.springframework.social.mailru.api.UsersOperations;
 import org.springframework.social.mailru.api.WallOperations;
 import org.springframework.social.oauth2.AbstractOAuth2ApiBinding;
+
 import java.util.LinkedList;
 import java.util.List;
 
@@ -37,11 +38,15 @@ public class MailruTemplate extends AbstractOAuth2ApiBinding implements Mailru {
 
     private WallOperations wallOperations;
 
-    private final String clientId;
+    private String clientId;
 
-    private final String clientSecret;
+    private String clientSecret;
 
-    private final String accessToken;
+    private String accessToken;
+
+    public MailruTemplate() {
+        initialize();
+    }
 
     public MailruTemplate(String clientId, String clientSecret, String accessToken) {
         super(accessToken);
@@ -60,11 +65,11 @@ public class MailruTemplate extends AbstractOAuth2ApiBinding implements Mailru {
     private void registerJsonModule() {
         List<HttpMessageConverter<?>> converters = getRestTemplate().getMessageConverters();
         for (HttpMessageConverter<?> converter : converters) {
-            if (converter instanceof MappingJacksonHttpMessageConverter) {
-                MappingJacksonHttpMessageConverter jsonConverter = (MappingJacksonHttpMessageConverter) converter;
+            if (converter instanceof MappingJackson2HttpMessageConverter) {
+                MappingJackson2HttpMessageConverter jsonConverter = (MappingJackson2HttpMessageConverter) converter;
 
                 List<MediaType> mTypes = new LinkedList<MediaType>(jsonConverter.getSupportedMediaTypes());
-                mTypes.add(new MediaType("text", "javascript", MappingJacksonHttpMessageConverter.DEFAULT_CHARSET));
+                mTypes.add(new MediaType("text", "javascript", MappingJackson2HttpMessageConverter.DEFAULT_CHARSET));
                 jsonConverter.setSupportedMediaTypes(mTypes);
 
                 ObjectMapper objectMapper = new ObjectMapper();
